Apply to MagicGirl

+++ autoupdater.sp
@@ -1,12 +1,12 @@
 #pragma semicolon 1
 #pragma newdecls required
 
-#include <cg_core>
+#include <MagicGirl.NET>
+#include <MagicGirl/vars>
 #include <system2>
 
-Handle g_hDatabase;
+Database g_hDatabase;
 bool testServer;
-char g_szMod[16];
 
 #include "updater/cfg.sp"   //自动更新cfg和翻译
 #include "updater/map.sp"   //自动添加地图到服务器
@@ -16,11 +16,11 @@ char g_szMod[16];
 
 public Plugin myinfo = 
 {
-    name        = "[CG] - Auto Updater",
+    name        = "[DARLING in the FRANXX] - Auto Updater",
     author      = "Kyle",
     description = "an auto update system",
     version     = "1.9.<commit_count>.<commit_branch> - <commit_date>",
-    url         = "http://steamcommunity.com/id/_xQy_/"
+    url         = "https://02.ditf.moe"
 };
 
 void OnDatabaseAvailable()
@@ -28,40 +28,21 @@ void OnDatabaseAvailable()
     if(testServer)
         return;
 
-    LogMessage("AutoUpdater is checking server[%d] now...", CG_GetServerId());
+    LogMessage("Auto-Updater is checking server [ %d ] now...", MG_Core_GetServerId());
 
-    switch(CG_GetServerId())
+    if(MG_Core_GetServerModId() == 199)
     {
-        case 1 : strcopy(g_szMod, 16, "ze");
-        case 2 : strcopy(g_szMod, 16, "ze");
-        case 3 : strcopy(g_szMod, 16, "ze");
-        case 4 : strcopy(g_szMod, 16, "ze");
-        case 5 : strcopy(g_szMod, 16, "tt");
-        case 6 : strcopy(g_szMod, 16, "tt");
-        case 7 : strcopy(g_szMod, 16, "mg");
-        case 8 : strcopy(g_szMod, 16, "jb");
-        case 9 : strcopy(g_szMod, 16, "jb");
-        case 11: strcopy(g_szMod, 16, "hg");
-        case 12: strcopy(g_szMod, 16, "ds");
-        case 15: strcopy(g_szMod, 16, "kz");
-        case 16: strcopy(g_szMod, 16, "kz");
-        case 19: strcopy(g_szMod, 16, "kz");
-        case 20: strcopy(g_szMod, 16, "kz");
-        default:
-        {
-            char m_szPath[128];
-            BuildPath(Path_SM, m_szPath, 128, "plugins/autoupdater.smx");
-            if(!FileExists(m_szPath) || !DeleteFile(m_szPath))
-                LogError("Delete autoupdater.smx failed.");
-            ServerCommand("sm plugins unload autoupdater.smx");
-            return;
-        }
+        char m_szPath[128];
+        BuildPath(Path_SM, m_szPath, 128, "plugins/autoupdater.smx");
+        if(!FileExists(m_szPath) || !DeleteFile(m_szPath))
+            LogError("Delete autoupdater.smx failed.");
+        ServerCommand("sm plugins unload autoupdater.smx");
+        return;
     }
     
     char m_szQuery[128];
-    FormatEx(m_szQuery, 128, "SELECT `map` FROM map_database WHERE `mod` = '%s'", g_szMod);
-    SQL_TQuery(g_hDatabase, SQLCallback_CheckMap, m_szQuery, 0);
-    
+    FormatEx(m_szQuery, 128, "SELECT `map` FROM dxg_mapdb WHERE `mod` = '%d'", MG_Core_GetServerModId());
+    g_hDatabase.Query(SQLCallback_CheckMap, m_szQuery);
     SMX_OnDatabaseAvailable();
 }
 
@@ -80,14 +61,22 @@ public void OnAllPluginsLoaded()
     MAP_OnAllPluginLoaded();
     ADD_OnAllPluginLoaded();
 }
+
+public void MG_MySQL_OnConnected(Database mysql)
+{
+    g_hDatabase = mysql;
+}
  
 void CheckDatabase()
 {
-    g_hDatabase = CG_DatabaseGetGames();
-    if(g_hDatabase == INVALID_HANDLE)
+    g_hDatabase = MG_MySQL_GetDatabase();
+    if(g_hDatabase == null)
+    {
         CreateTimer(5.0, Timer_Reconnect);
-    else
-        OnDatabaseAvailable();
+        return;
+    }
+
+    OnDatabaseAvailable();
 }
 
 public Action Timer_Reconnect(Handle timer)
@@ -108,12 +97,12 @@ public void OnMapVoteEnd(const char[] map)
 {
     if(StrContains(map, "extend", false) != -1)
         return;
-    
+
     if(StrContains(map, "change", false) != -1)
         return;
 
     strcopy(nextMap, 128, map);
-    
+
     NAV_CheckMapNav(map);
 
     CFG_GetTrans();
